# backend/api/swagger.json
{
    "swagger" : "2.0",
    "info" : {
      "description" : "Cryptotrades API Doc",
      "version" : "v1",
      "title" : "Cryptotrades"
    },
    "host" : "localhost:3000",
    "basePath" : "/",
    "schemes" : [ "http", "https" ],
...
    "paths" : {
...
          "/settings/setoptions" : {
            "post" : {
              "tags" : [ "admin" ],
              "summary" : "Set Options",
              "description" : "",
              "operationId" : "setoptions",
              "consumes" : [ "application/x-www-form-urlencoded" ],
              "produces" : [ "application/json" ],
              "parameters" : [{
                "name" : "authorization",
                "in" : "header",
                "description" : "",
                "required" : true,
                "type" : "string",
                "default" : ""
              },{
                "name" : "name",
                "in" : "formData",
                "description" : "Enter Name",
                "required" : true,
                "type" : "string"
              },{
                "name" : "value",
                "in" : "formData",
                "description" : "Enter Value",
                "required" : true,
                "type" : "string"
              }],
              "responses" : {
                "405" : {
                  "description" : "Invalid input"
                }
              }
            }
          }
        }
    }